Part 1/Exercise 1.15/requirements.txt
requests==2.25.1
pillow
